+++ src/decompx.jl
@@ -5,7 +5,7 @@ Decomposition of a matrix by orthogonal projection on factors.
 * `X` :  X-data (n, p) to decompose.
 * `f` : A formula that defines the factor(s) on which is(are) done the decomposition.
     See the syntax in the examples below.
-* `dat` (n, q) : DataFrame containing the factor(s) specified in `f`. 
+* `dat` (n, q) : Dataframe containing the factor(s) specified in `f`. 
 
 ## References
 Bertinetto, C., Engel, J., Jansen, J., 2020. ANOVA simultaneous component analysis: A tutorial review.