+++ src/manova.jl
@@ -1,19 +1,19 @@
 """
-    manova(; digits = 4)
-MANOVA with approximated F tests.
-* `b` : Vector (p) of the coefficients of the model.
-* `L` : Matrix (m, p) such as `L` * `b` gives the linear combination(s) of the coefficients 
-    to be tested.
-* `varb` : Variance-covariance matrix (p, p) of `b`.
+    manova(Y, f::StatsModels.FormulaTerm, dat::DataFrame; test = :pillai, digits = 4)
+MANOVA.
+* `Y` : Y-data (n, p) representing the response variables.
+* `f` : A formula that defines the tested factor(s). See the syntax in the examples below.
+* `dat` (n, q) : Dataframe containing the factor(s) specified in `f`. 
 Keyword arguments:
+* `test` : Type of statistic used for the test. Possible values are: `:wilks`, `:pillai` (default), 
+    `:hotelling`, or `:roy`.
 * `digits` : Nb. digits for the outputs.
 
-The function tests hypothesis H0: `L` * `b` = `h0`, with either 
-* a Chi-squared Wald test (with dfs = m)
-* or, if `defden` is given, a F test (with dfs {m, `defden`}).
-
-Both tests assume that `b` is Gaussian.  Compared to the F test, the Wald test neglects the uncertainty 
-affecting the estimate of the dispersion parameter of the model (e.g., 'sigma2' in MLRs). 
+The function returns approximated F tests for one of the the following statistics (argument `test`):
+* Wilks’ lambda
+* Pillai’s trace
+* Hotelling-Lawley trace
+* Roy’s maximum root
 
 ## References
 https://documentation.sas.com/doc/en/statug/15.2/statug_introreg_sect038.htm#statug_introreg001918
@@ -23,6 +23,7 @@ https://documentation.sas.com/doc/en/statug/15.2/statug_introreg_sect038.htm#sta
 ```
 """
 function manova(Y, f::StatsModels.FormulaTerm, dat::DataFrame; test = :pillai, digits = 4)
+    @assert in([:wilks; :pillai; :hotelling; :roy])(test) "Wrong value for argument 'test'." 
     Y = ensure_mat(Y)
     Q = eltype(Y)
     res = decompx(Y, f, dat)
@@ -37,7 +38,7 @@ function manova(Y, f::StatsModels.FormulaTerm, dat::DataFrame; test = :pillai, d
     val = list(Q, length(L))
     F = copy(val)
     dfnum = list(Int, length(L))
-    defden = list(Int, length(L))
+    dfden = list(Int, length(L))
     pval = copy(val)
     for i in eachindex(L)
         LB = L[i] * B
@@ -50,16 +51,50 @@ function manova(Y, f::StatsModels.FormulaTerm, dat::DataFrame; test = :pillai, d
         s = min(p, q)
         m = (abs(p - q) - 1) / 2
         n = (nu - p - 1) / 2
-        if test == :pillai
-            F[i] = (2 * n + s + 1) / (2 * m + s + 1) * val[i] / (s - val[i]) 
+        if test == :wilks
+            c = p^2 + q^2 - 5
+            if c > 0
+                t = sqrt(((p * q)^2 - 4) / c)
+            else
+                t = 1
+            end
+            r = nu - (p - q + 1) / 2
+            u = (p * q - 2) / 4 
+            a = r * t - 2 * u
+            b = p * q
+            F[i] = a / b * (1 - val[i]^(1 / t)) / val[i]^(1 / t) 
+            dfnum[i] = b 
+            dfden[i] = a
+        elseif test == :pillai
+            a = 2 * n + s + 1
+            b = 2 * m + s + 1
+            F[i] = a / b * val[i] / (s - val[i]) 
+            dfnum[i] = s * b
+            dfden[i] = s * a
+        elseif test == :hotelling
+            F[i] = 2 * (s * n + 1) * val[i] / (s^2 * (2 * m + s + 1))
             dfnum[i] = s * (2 * m + s + 1)
-            defden[i] = s * (2 * n + s + 1)
+            dfden[i] = 2 * (s * n + 1)
+            ## This version of the function follows the R's choice to not use
+            ## the following variant when n > 0 (used in SAS, see ref):
+            ## b = (p + 2 * n) * (q + 2 * n) / (2 * (2 * n + 1) * (n - 1))
+            ## c = (2 + (p * q + 2) / (b - 1)) / (2 * n)
+            ## F[i] = (val[i] / c) * ((4 + (p * q + 2) / (b - 1)) / (p * q))
+            ## dfnum[i] = p * q
+            ## dfden[i] = 4 + (p * q + 2) / (b - 1)
+            ## End
+        elseif test == :roy
+            r = max(p, q)
+            a = nu - r + q
+            F[i] = a / r * val[i]
+            dfnum[i] = r
+            dfden[i] = a
         end
-        d = Distributions.FDist(dfnum[i], defden[i])
+        d = Distributions.FDist(dfnum[i], dfden[i])
         pval[i] = Distributions.ccdf(d, F[i])
-    end
+    end        
     pval .= round.(pval; digits)
     nam = collect(@names res.fit)[2:end]
-    DataFrame(:term => nam, test => val, :approxF => F, :dfnum => dfnum, :defden => defden, 
-        :pval => pval)
+    df = res.df.dffit[2:end]
+    DataFrame(:term => nam, :df => df, test => val, :approxF => F, :dfnum => dfnum, :dfden => dfden, :pval => pval)
 end